+++ CHANGES.md
@@ -6,6 +6,29 @@ This release supports [version
 
 ## New Features
 
+* SAWScript now supports optional named arguments.
+  These can be passed anywhere in a function application using the
+  syntax `name=val`.
+  If `val` requires parentheses you can also enclose the name in the
+  parentheses as well.
+  The syntax in a type signature is `name?type`.
+  The parameter syntax in a function declaration is either `name?=val`
+  or `name@alt?=val`, where `val` is the default value to use when
+  the caller doesn't provide one.
+  By default the `name` is both the external call name and the internal
+  variable name for the parameter.
+  If `alt` is present, it can be an identifier or a pattern, and specifies
+  a different name or names for internal use.
+  The alternate name, or parts of it, can be set to `_` if unused.
+
+* The builtin `str_concats` now accepts an optional argument `sep`
+  to use as a separator between the given strings.
+  For example, `str_concats ["a", "b"]` will produce `"ab"` as before,
+  but `str_concats sep=", " ["a", "b"]` produces `"a, b"`.
+
+* The SAWScript typechecker will now (sometimes, at least) guess when
+  you forgot a semicolon.
+
 * The new `subproof` builtin is similar to "bullets" in Rocq:
   When there are multiple remaining proof goals, it runs the given
   inner proof script on the first goal alone, and then checks that the
@@ -83,6 +106,10 @@ This release supports [version
   Use `admit` instead.
 
 ## Other changes
+
+* The way the `:search` REPL command matches function types has been adjusted.
+  It should be more predictable and more useful now.
+  Feedback is welcome.
 
 * `crux-mir-comp` has been unbundled from the SAW release distribution
   in preparation for creating [a standalone Crux release](https://github.com/GaloisInc/crux).